structure improvement - part 1
